requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-jsonfield==1.4.1
django-tenants==3.4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
